+++ src/utils/schemaGetters.js
@@ -1,23 +1,32 @@
 import * as R from 'ramda'
+import { titleize, humanize } from '../Utils'
+import pluralize from "pluralize"
 
-export const getFieldLabel = ({ schema, modelName, fieldName, data = {}, customProps }) => {
-  const displayName = R.pathOr('No Name Found', [modelName, 'fields', fieldName, 'displayName'], schema)
+// This is an example of our naming issue, node and data should not both be passed here
+export const getFieldLabel = ({ schema, modelName, fieldName, node, data, customProps }) => {
+  const displayName = R.pathOr(
+    humanize(fieldName),
+    [modelName, 'fields', fieldName, 'displayName'],
+    schema
+  )
   if (R.type(displayName) === 'Function') {
-    return displayName({ schema, modelName, data, customProps })
+    return displayName({ schema, modelName, node, data, customProps })
   }
   return displayName
 }
 
-export const getModelLabel = ({ schema, modelName, data, customProps }) => {
-  const displayName = R.pathOr('No Name Found', [modelName, 'displayName'], schema)
+export const getModelLabel = ({ schema, modelName, node, data, customProps }) => {
+  const defaultValue = titleize(humanize(modelName))
+  const displayName = R.pathOr(defaultValue, [modelName, 'displayName'], schema)
   if (R.type(displayName) === 'Function') {
-    return displayName({ schema, modelName, data, customProps })
+    return displayName({ schema, modelName, node, data, customProps })
   }
   return displayName
 }
 
 export const getModelLabelPlural = ({ schema, modelName, data, user, customProps }) => {
-  const displayName = R.pathOr('No Name Found', [modelName, 'displayNamePlural'], schema)
+  const defaultValue = pluralize(titleize(modelName))
+  const displayName = R.pathOr(defaultValue, [modelName, 'displayNamePlural'], schema)
   if (R.type(displayName) === 'Function') {
     return displayName({ schema, modelName, data, user, customProps })
   }
@@ -50,10 +59,26 @@ export const getField = (schema, modelName, fieldName) => (
   )(schema, modelName)
 )
 
-const getShownFields = ({ schema, modelName, type, node = {}, data, user, customProps }) => {
+const getShownFields = ({ schema, modelName, type, node, data, user, customProps }) => {
   const fieldOrder = R.prop('fieldOrder', getModel(schema, modelName))
   return R.filter(fieldName => {
-    let show = R.prop(type, getField(schema, modelName, fieldName))
+    let show
+    switch (type) {
+      case 'showCreate':
+      case 'showDetail':
+        show = R.propOr(
+          !R.equals('id', fieldName),
+          type,
+          getField(schema, modelName, fieldName)
+        )
+        break
+      case 'showIndex':
+      case 'showTooltip':
+        show = R.propOr(false, type, getField(schema, modelName, fieldName))
+        break
+      default:
+        show = R.prop(type, getField(schema, modelName, fieldName))
+    }
     if (R.type(show) === 'Function') {
       show = show({
         schema, modelName, fieldName, node, data, user, customProps
@@ -79,7 +104,7 @@ export const getCreateFields = ({ schema, modelName, user, customProps }) => {
 }
 
 export const getHasIndex = (schema, modelName) => {
-  return R.prop('hasIndex', getModel(schema, modelName))
+  return R.propOr(true, 'hasIndex', getModel(schema, modelName))
 }
 
 export const getDetailFields = ({ schema, modelName, node, customProps }) => {
@@ -111,4 +136,8 @@ export const getEnumChoices = (schema, modelName, fieldName) => {
 
 export const getEnumChoiceOrder = (schema, modelName, fieldName) => {
   return R.prop('choiceOrder', getField(schema, modelName, fieldName))
+}
+
+export const getFieldConditions = (schema, modelName, fieldName) => {
+  return R.prop('displayConditions', getField(schema, modelName, fieldName))
 }